renamed file and changed testes
